Initial Commit
This establishes a working dns configuration with good config
directory support based on DNSMasq. It also supports dns over tls
with an older stubby container.

There is a lot of room for improvement here, but I wanted to get
something running and these are the tools I knew.

Look at PiHole, unbound for future iterations.